+++ packages/streams/src/methods/create.ts
@@ -1,11 +1,8 @@
-import { BASE_URL } from '../MoralisStreams';
-import { EndpointResolver } from '@moralisweb3/api-utils';
-import MoralisCore, { MoralisStreamError, StreamErrorCode } from '@moralisweb3/core';
+import { Endpoints } from '@moralisweb3/api-utils';
 import { EvmAddressish, EvmChainish } from '@moralisweb3/evm-utils';
 import { createStreamEvm } from '../resolvers';
-export enum StreamNetwork {
-  EVM = 'evm',
-}
+import { IncorrectNetworkError } from '../utils/IncorrectNetworkError';
+import { StreamNetwork } from '../utils/StreamNetwork';
 
 export interface CreateStreamEvmOptions {
   network: 'evm';
@@ -28,20 +25,15 @@ export interface CreateStreamEvmOptions {
 
 export type CreateStreamOptions = CreateStreamEvmOptions;
 
-const makeCreateStreamEvm = (core: MoralisCore, { network, ...options }: CreateStreamEvmOptions) => {
-  return EndpointResolver.create(core, BASE_URL, createStreamEvm).fetch(options);
-};
+export const makeCreateStream = (endpoints: Endpoints) => {
+  const evmFetcher = endpoints.createFetcher(createStreamEvm);
 
-export const createStream = (core: MoralisCore) => (options: CreateStreamOptions) => {
-  switch (options.network) {
-    case StreamNetwork.EVM:
-      return makeCreateStreamEvm(core, options);
-    default:
-      throw new MoralisStreamError({
-        code: StreamErrorCode.INCORRECT_NETWORK,
-        message: `Incorrect network provided. Got "${options.network}", Valid values are: ${Object.values(StreamNetwork)
-          .map((value) => `"${value}"`)
-          .join(', ')}`,
-      });
-  }
+  return ({ network, ...options }: CreateStreamOptions) => {
+    switch (network) {
+      case StreamNetwork.EVM:
+        return evmFetcher({ ...options });
+      default:
+        throw new IncorrectNetworkError(network);
+    }
+  };
 };