+++ packages/streams/src/methods/delete.ts
@@ -1,10 +1,7 @@
-import { BASE_URL } from '../MoralisStreams';
-import { EndpointResolver } from '@moralisweb3/api-utils';
-import MoralisCore, { MoralisStreamError, StreamErrorCode } from '@moralisweb3/core';
+import { Endpoints } from '@moralisweb3/api-utils';
 import { deleteStreamEvm } from '../resolvers';
-export enum StreamNetwork {
-  EVM = 'evm',
-}
+import { StreamNetwork } from '../utils/StreamNetwork';
+import { IncorrectNetworkError } from '../utils/IncorrectNetworkError';
 
 export interface DeleteStreamEvmOptions {
   network: 'evm';
@@ -13,20 +10,15 @@ export interface DeleteStreamEvmOptions {
 
 export type DeleteStreamOptions = DeleteStreamEvmOptions;
 
-const makeDeleteStreamEvm = (core: MoralisCore, { network, ...options }: DeleteStreamEvmOptions) => {
-  return EndpointResolver.create(core, BASE_URL, deleteStreamEvm).fetch(options);
-};
+export const makeDeleteStream = (endpoints: Endpoints) => {
+  const evmFetcher = endpoints.createFetcher(deleteStreamEvm);
 
-export const deleteStream = (core: MoralisCore) => (options: DeleteStreamOptions) => {
-  switch (options.network) {
-    case StreamNetwork.EVM:
-      return makeDeleteStreamEvm(core, options);
-    default:
-      throw new MoralisStreamError({
-        code: StreamErrorCode.INCORRECT_NETWORK,
-        message: `Incorrect network provided. Got "${options.network}", Valid values are: ${Object.values(StreamNetwork)
-          .map((value) => `"${value}"`)
-          .join(', ')}`,
-      });
-  }
+  return ({ network, ...options }: DeleteStreamOptions) => {
+    switch (network) {
+      case StreamNetwork.EVM:
+        return evmFetcher({ ...options });
+      default:
+        throw new IncorrectNetworkError(network);
+    }
+  };
 };