+++ packages/streams/src/resolvers/evmStreams/deleteAddressEvm.ts
@@ -1,14 +1,17 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 import { toCamelCase } from '@moralisweb3/core';
-import { EvmAddress, EvmChain } from '@moralisweb3/evm-utils';
-import { operations } from '../generated/types';
+import { EvmAddress } from '@moralisweb3/evm-utils';
+import { operations } from '../../generated/types';
 
-const name = 'DeleteStream';
+const name = 'DeleteAddressFromStream';
 
 type Name = typeof name;
 type PathParams = operations[Name]['parameters']['path'];
-type ApiParams = PathParams;
+type BodyParams = operations[Name]['requestBody']['content']['application/json'];
+type ApiParams = PathParams & BodyParams;
+export type DeleteAddressEvmParams = ApiParams;
 const method = 'delete';
+const bodyParams = ['address'] as const;
 
 type ApiResult = operations[Name]['responses']['200']['content']['application/json'];
 
@@ -18,23 +21,20 @@ const apiToResult = (apiData: ApiResult) => {
   return {
     ...data,
     address: data.address ? EvmAddress.create(data.address) : undefined,
-    tokenAddress: data.tokenAddress ? EvmAddress.create(data.tokenAddress) : undefined,
-    chains: data.chainIds.map((chainId) => EvmChain.create(chainId)),
   };
 };
 
-export const deleteStreamEvm = createEndpointFactory(() =>
+export const deleteAddressEvm = createEndpointFactory(() =>
   createEndpoint({
     name,
-    getUrl: (params: ApiParams) => `/streams/evm/${params.id}`,
+    getUrl: (params: DeleteAddressEvmParams) => `/streams/evm/${params.id}/address`,
     apiToResult,
     resultToJson: (data) => ({
       ...data,
       address: data.address ? data.address.format() : undefined,
-      tokenAddress: data.tokenAddress ? data.tokenAddress.format() : undefined,
-      chainIds: data.chains.map((chain) => chain.format()),
     }),
-    parseParams: (params: ApiParams): ApiParams => params,
+    parseParams: (params: DeleteAddressEvmParams): ApiParams => params,
     method,
+    bodyParams,
   }),
 );