+++ packages/streams/src/resolvers/history/getHistory.ts
@@ -1,18 +1,18 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
-import { operations } from '../generated/types';
+import { operations } from '../../generated/types';
 
-const name = 'ReplayHistory';
+const name = 'GetHistory';
 
 type Name = typeof name;
 type ApiResult = operations[Name]['responses']['200']['content']['application/json'];
-type PathParams = operations[Name]['parameters']['path'];
-type ApiParams = PathParams;
+type QueryParams = operations[Name]['parameters']['query'];
+type ApiParams = QueryParams;
 type Params = ApiParams;
 
-export const replayHistory = createEndpointFactory(() =>
+export const getHistory = createEndpointFactory(() =>
   createEndpoint({
     name,
-    getUrl: (id: Params) => `/history/replay/${id}`,
+    getUrl: () => `/history`,
     apiToResult: (data: ApiResult) => data,
     resultToJson: (data) => data,
     parseParams: (params: Params): ApiParams => params,