+++ packages/solApi/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @moralisweb3/sol-api
 
+## 2.6.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.6.0
+  - @moralisweb3/api-utils@2.6.0
+  - @moralisweb3/sol-utils@2.6.0
+
 ## 2.5.8
 
 ### Patch Changes